Move path_matcher to independent build target
