+++ contrib/endpoints/src/api_manager/context/BUILD
@@ -35,6 +35,7 @@ cc_library(
     }),
     deps = [
         "//contrib/endpoints/src/api_manager:http_template",
+        "//contrib/endpoints/src/api_manager:path_matcher",
         "//contrib/endpoints/src/api_manager:impl_headers",
         "//contrib/endpoints/src/api_manager:server_config_proto",
         "//contrib/endpoints/src/api_manager/auth",
